Moved most_focus method into it's own helper file, also moved top_three_focus and profile_exists? and declared them private
